+++ projects/signal-plus/src/lib/reactive-queries/interfaces/infinite-query-result.interface.ts
@@ -1,0 +1,1 @@
+export type { InfiniteQueryResult } from '../../models/query.model';